add password change tests
