+++ Pipfile
@@ -12,7 +12,7 @@ dj-database-url = "*"
 psycopg2-binary = "*"
 autopep8 = "*"
 django-crispy-forms = "*"
-django-heroku= "*"
+django-heroku = "==0.3.1"
 
 [requires]
 python_version = "3.7"
