+++ pytest.ini
@@ -1,6 +1,5 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = config.default
-addopts = --cov-report html --cov-report term-missing --cov=. 
-python_files = tests.py test_*.py *_tests.py
+addopts = --cov-report html --cov-report term-missing --cov=app 
 
 exclude = tests/*
